refactor: getPref.ts
