+++ packages/supi/src/utils/getPref.ts
@@ -1,5 +1,87 @@
 import PnpmError from '@pnpm/error'
+import { ResolvedDirectDependency } from '@pnpm/resolve-dependencies'
 import versionSelectorType = require('version-selector-type')
+import { ImporterToUpdate } from '../install'
+import save, { PackageSpecObject } from '../save'
+
+export async function updateImporterManifest (
+  importer: ImporterToUpdate,
+  opts: {
+    directDependencies: ResolvedDirectDependency[],
+    saveWorkspaceProtocol: boolean,
+  },
+) {
+  if (!importer.manifest) {
+    throw new Error('Cannot save because no package.json found')
+  }
+  const specsToUpsert = opts.directDependencies.map((rdd) => resolvedDirectDepToSpecObject(rdd, importer, opts))
+  for (const pkgToInstall of importer.wantedDeps) {
+    if (pkgToInstall.alias && !specsToUpsert.some(({ alias }) => alias === pkgToInstall.alias)) {
+      specsToUpsert.push({
+        alias: pkgToInstall.alias,
+        peer: importer['peer'],
+        saveType: importer['targetDependenciesField'],
+      })
+    }
+  }
+  return save(
+    importer.rootDir,
+    importer.manifest,
+    specsToUpsert,
+    { dryRun: true },
+  )
+}
+
+function resolvedDirectDepToSpecObject (
+  {
+    alias,
+    name,
+    normalizedPref,
+    resolution,
+    specRaw,
+    version,
+  }: ResolvedDirectDependency,
+  importer: ImporterToUpdate,
+  opts: {
+    saveWorkspaceProtocol: boolean,
+  }
+): PackageSpecObject {
+  let pref!: string
+  const isNew = importer.newPkgRawSpecs.includes(specRaw)
+  if (normalizedPref) {
+    pref = normalizedPref
+  } else {
+    if (isNew) {
+      pref = getPrefPreferSpecifiedSpec({
+        alias,
+        name,
+        pinnedVersion: importer['pinnedVersion'],
+        specRaw,
+        version,
+      })
+    } else {
+      pref = getPrefPreferSpecifiedExoticSpec({
+        alias,
+        name,
+        specRaw,
+        version,
+      })
+    }
+    if (
+      resolution.type === 'directory' &&
+      opts.saveWorkspaceProtocol &&
+      !pref.startsWith('workspace:')
+    ) {
+      pref = `workspace:${pref}`
+    }
+  }
+  return {
+    alias,
+    peer: importer['peer'],
+    pref,
+    saveType: isNew ? importer['targetDependenciesField'] : undefined,
+  }
+}
 
 export type PinnedVersion = 'major' | 'minor' | 'patch'
 
@@ -17,40 +99,40 @@ export default function getPref (
   return `${prefix}${createVersionSpec(version, opts.pinnedVersion)}`
 }
 
-export function getPrefPreferSpecifiedSpec (
+function getPrefPreferSpecifiedSpec (
   opts: {
     alias: string,
     name: string
     version: string,
-    rawSpec: string,
+    specRaw: string,
     pinnedVersion?: PinnedVersion,
   },
  ) {
   const prefix = getPrefix(opts.alias, opts.name)
-  if (opts.rawSpec?.startsWith(`${opts.alias}@${prefix}`)) {
-    const selector = versionSelectorType(opts.rawSpec.substr(`${opts.alias}@${prefix}`.length))
+  if (opts.specRaw?.startsWith(`${opts.alias}@${prefix}`)) {
+    const selector = versionSelectorType(opts.specRaw.substr(`${opts.alias}@${prefix}`.length))
     if (selector && (selector.type === 'version' || selector.type === 'range')) {
-      return opts.rawSpec.substr(opts.alias.length + 1)
+      return opts.specRaw.substr(opts.alias.length + 1)
     }
   }
   return `${prefix}${createVersionSpec(opts.version, opts.pinnedVersion)}`
 }
 
-export function getPrefPreferSpecifiedExoticSpec (
+function getPrefPreferSpecifiedExoticSpec (
   opts: {
     alias: string,
     name: string
     version: string,
-    rawSpec: string,
+    specRaw: string,
     pinnedVersion?: PinnedVersion,
   },
 ) {
   const prefix = getPrefix(opts.alias, opts.name)
-  if (opts.rawSpec?.startsWith(`${opts.alias}@${prefix}`)) {
-    const specWithoutName = opts.rawSpec.substr(`${opts.alias}@${prefix}`.length)
+  if (opts.specRaw?.startsWith(`${opts.alias}@${prefix}`)) {
+    const specWithoutName = opts.specRaw.substr(`${opts.alias}@${prefix}`.length)
     const selector = versionSelectorType(specWithoutName)
     if (!(selector && (selector.type === 'version' || selector.type === 'range'))) {
-      return opts.rawSpec.substr(opts.alias.length + 1)
+      return opts.specRaw.substr(opts.alias.length + 1)
     }
     if (!opts.pinnedVersion) {
       switch (selector.type) {
